Reorganized sub components into a its own folder. Create a page section component called OceanSection that uses those sub components
