+++ src/components/ocean/OceanComponents/Seaweed.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { classNames } from "../../utils/string";
-import Seaweeded from "../../assets/ocean/Seaweeded.svg";
+import { classNames } from "../../../utils/string";
+import Seaweeded from "../../../assets/ocean/Seaweeded.svg";
 
 const Seaweed = ({ className }: { className?: string }) => {
   return (
